Fix: Change lightTypes from int to float for shader compatibility
IGraphicsBackend::setShaderParameterArray only supports f32, Vec2f, and Vec3f
arrays, not int arrays. Changed lightTypes to f32 throughout:

- DynamicLightingEffect.cpp: Changed types vector from i32 to f32
- dynamic_lighting.frag: Changed uniform lightTypes from int to float
- Comparisons still work (implicit conversion in GLSL)

+++ client/assets/shaders/dynamic_lighting.frag
@@ -12,7 +12,7 @@ uniform vec2 lightPositions[MAX_LIGHTS];    // Positions monde (pixels)
 uniform vec3 lightColors[MAX_LIGHTS];       // Couleurs RGB (0-1)
 uniform float lightRadius[MAX_LIGHTS];      // Rayons monde (pixels)
 uniform float lightIntensity[MAX_LIGHTS];   // Intensité de teinte (0-1)
-uniform int lightTypes[MAX_LIGHTS];         // Type de lumière (0=Point, 1=Directional, 2=Spot)
+uniform float lightTypes[MAX_LIGHTS];       // Type de lumière (0=Point, 1=Directional, 2=Spot) - float for compatibility
 uniform vec2 lightDirections[MAX_LIGHTS];   // Direction pour Directional/Spot
 uniform float lightAngles[MAX_LIGHTS];      // Angle du cône pour Spot (degrés)
 uniform float ambientDarkness;              // Obscurité ambiante (0-1)
